feat(chart): split config and context into brain primitive
Move ChartConfig, ChartContext, CHART_CONTEXT token, resolveCssVar, and THEMES
from @spartan-ng/helm/chart to @spartan-ng/brain/chart.

Helm re-exports all symbols from brain for backward compatibility.
Users can now import from either package.
