+++ libs/brain/chart/src/lib/brn-chart.token.ts
@@ -1,0 +1,7 @@
+import { InjectionToken, type Signal } from '@angular/core';
+
+export type BrnChartState = {
+	activePoint: Signal<{ name: string; value: number; color: string; seriesName?: string } | null>;
+};
+
+export const BRN_CHART_STATE = new InjectionToken<BrnChartState>('BrnChartState');
