+++ libs/helm/chart/src/lib/chart-config.ts
@@ -1,9 +1,1 @@
-export const THEMES = { light: '', dark: '.dark' } as const;
-
-export type ChartConfig = Record<
-	string,
-	{
-		label?: string;
-		icon?: string;
-	} & ({ color?: string; theme?: never } | { color?: never; theme: Record<keyof typeof THEMES, string> })
->;
+export { THEMES, type ChartConfig } from '@spartan-ng/brain/chart';
