+++ tsconfig.base.json
@@ -42,6 +42,7 @@
 			"@spartan-ng/brain/avatar": ["libs/brain/avatar/src/index.ts"],
 			"@spartan-ng/brain/button": ["libs/brain/button/src/index.ts"],
 			"@spartan-ng/brain/calendar": ["libs/brain/calendar/src/index.ts"],
+			"@spartan-ng/brain/chart": ["libs/brain/chart/src/index.ts"],
 			"@spartan-ng/brain/checkbox": ["libs/brain/checkbox/src/index.ts"],
 			"@spartan-ng/brain/collapsible": ["libs/brain/collapsible/src/index.ts"],
 			"@spartan-ng/brain/combobox": ["libs/brain/combobox/src/index.ts"],
